update v7.5.0 components;

+++ components/aos/src/adapter/rhino/hook_impl.c
@@ -30,14 +30,14 @@ void krhino_init_hook(void)
     soc_hw_timer_init();
 #endif
 
-    uint32_t h_end = (uint32_t)(&__heap_end);
-    uint32_t h_start = (uint32_t)(&__heap_start);
+    size_t h_end = (size_t)(&__heap_end);
+    size_t h_start = (size_t)(&__heap_start);
 
     if (h_start >= h_end) {
         k_err_proc(RHINO_MM_POOL_SIZE_ERR);
     }
     /* auto define heap size */
-    g_mm_region[0].len = (uint32_t)(&__heap_end) - (uint32_t)(&__heap_start);
+    g_mm_region[0].len = (size_t)(&__heap_end) - (size_t)(&__heap_start);
 }
 
 __attribute__((weak)) void krhino_start_hook(void)
